+++ src/CoVoX/API/Modules/Translating/Translator.cs
@@ -63,16 +63,5 @@ namespace API.Modules
                 Log.Debug($"Stopped continuous recognition {translationRecognizer.SpeechRecognitionLanguage}");
             }
         }
-
-        public class TextRecognizedArgs : EventArgs
-        {
-            public string Text { get; }
-
-            public TextRecognizedArgs(string translatedText)
-            {
-                Text = new string(translatedText.Where(c => !char.IsPunctuation(c)).ToArray());
-                Log.Debug($"Detected text: {translatedText}");
-            }
-        }
     }
 }
