+++ src/CoVoX/CoVoX.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API", "API\API.csproj", "{E
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo", "Demo\Demo.csproj", "{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LigthSwitchDemo", "LigthSwitchDemo\LigthSwitchDemo.csproj", "{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightSwitchDemo", "LightSwitchDemo\LightSwitchDemo.csproj", "{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
