+++ src/CoVoX/LightSwitchDemo/App.xaml
