+++ src/CoVoX/LightSwitchDemo/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows;
+
+namespace LigthSwitchDemo
+{
+    /// <summary>
+    /// Interaction logic for App.xaml
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+}
