+++ src/CoVoX/LightSwitchDemo/AssemblyInfo.cs
