+++ src/CoVoX/LightSwitchDemo/LightSwitchDemo.csproj
@@ -4,6 +4,8 @@
     <OutputType>WinExe</OutputType>
     <TargetFramework>net5.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
+    <AssemblyName>LightSwitchDemo</AssemblyName>
+    <RootNamespace>LightSwitchDemo</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
