+++ src/CoVoX/LightSwitchDemo/MainWindow.xaml
