+++ src/CoVoX/LightSwitchDemo/appsettings.json
