+++ src/CoVoX/LightSwitchDemo/dummysecrets.json
